requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Cors==5.0.0
Flask-MySQLdb==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysqlclient==2.2.6
numpy==2.2.0
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.0.0
pydantic==2.10.3
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
scikit-video==1.1.11
scipy==1.14.1
six==1.17.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.1.3
